feat(logo): add new logo components for PNPM, Shadcn, Tailwind, Storybook, Vite, Nx, and React
