+++ libs/shadcnui-blocks/src/lib/components/logo/logos/ShadcnLogo.tsx
@@ -1,0 +1,36 @@
+import { SVGProps } from 'react';
+
+const ShadcnLogo = (props: SVGProps<SVGSVGElement>) => (
+  <svg
+    data-testid="shadcn-logo"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 256 256"
+    {...props}
+  >
+    <rect width="256" height="256" fill="none" />
+    <line
+      x1="208"
+      y1="128"
+      x2="128"
+      y2="208"
+      fill="none"
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="32"
+    />
+    <line
+      x1="192"
+      y1="40"
+      x2="40"
+      y2="192"
+      fill="none"
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="32"
+    />
+  </svg>
+);
+
+export { ShadcnLogo };
